+++ python/aswfdocker/tests/test_index.py
@@ -24,7 +24,7 @@ class TestIndex(unittest.TestCase):
         self.assertEqual(images[0], "baseos-gl-conan")
         versions = list(self.index.iter_versions(constants.ImageType.IMAGE, images[0]))
         self.assertGreaterEqual(len(versions), 1)
-        self.assertTrue(versions[0].startswith("4"))
+        self.assertTrue(versions[0].startswith("3"))
 
     def test_version_info(self):
         vi = self.index.version_info("2019")
@@ -34,5 +34,5 @@ class TestIndex(unittest.TestCase):
     def test_group_from_image(self):
         self.assertEqual(
             self.index.get_group_from_image(constants.ImageType.PACKAGE, "clang"),
-            "common",
+            "common-2",
         )